# Store-Gateway: Reduce generated garbage when loading chunks from bucket block