… pkg/storegateway/pool_test.go