<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

## What this PR does

Motivation of this PR is to reduce total amount of generated garbage when populating chunks from a bucket block. 

### The profile

Below there's a screenshot of a `storegateway` memory profile corresponding to `cortex-prod-01` cell.

![Screenshot 2022-10-14 at 11 55 31](https://user-images.githubusercontent.com/888899/195823978-2e6a1aab-ca41-41e6-8d5e-7061f5c34d57.png)

According to this, roughly `26%` of total allocated objects are generated while invoking `loadChunks`,  and about the `22%` is distributed between the functions `populateChunk` and `loadChunk` itself.

[store-gateway.cortex-prod-01.mem.profile.zip](https://github.com/grafana/mimir/files/9785502/store-gateway.cortex-prod-01.mem.profile.zip)

### The changes

When listing `populateChunk` we observe the following output:

```
(pprof) list populateChunk
Total: 16800841
ROUTINE ======================== github.com/grafana/mimir/pkg/storegateway.populateChunk in /backend-enterprise/vendor/github.com/grafana/mimir/pkg/storegateway/bucket.go
   1884218    1898590 (flat, cum) 11.30% of Total
         .          .    792:	return ps, stats
         .          .    793:}
         .          .    794:
         .          .    795:func populateChunk(out *storepb.AggrChunk, in chunkenc.Chunk, aggrs []storepb.Aggr, save func([]byte) ([]byte, error)) error {
         .          .    796:	if in.Encoding() == chunkenc.EncXOR {
         .      14372    797:		b, err := save(in.Bytes())
         .          .    798:		if err != nil {
         .          .    799:			return err
         .          .    800:		}
   1884218    1884218    801:		out.Raw = &storepb.Chunk{Type: storepb.Chunk_XOR, Data: b}
         .          .    802:		return nil
         .          .    803:	}
         .          .    804:	return errors.Errorf("unsupported chunk encoding %d", in.Encoding())
         .          .    805:}
         .          .    806:
(pprof)
```

Here we can see that most of allocated objects in this function correspond to the `storepb.Chunk` object type.

To mitigitate this, as part of this PR we've included an extra pool, shared across block readers, to recycle this type of objects.

Much more interesting is the content of the listing for the `loadChunks` function:

```
(pprof) list loadChunks
Total: 16800841
ROUTINE ======================== github.com/grafana/mimir/pkg/storegateway.(*bucketChunkReader).loadChunks in /backend-enterprise/vendor/github.com/grafana/mimir/pkg/storegateway/bucket.go
   1792365    4389100 (flat, cum) 26.12% of Total
[...]
         .          .   2595:		cb := buf[:chunkLen]
         .          .   2596:		n, err = io.ReadFull(bufReader, cb)
         .          .   2597:		readOffset += n
         .          .   2598:		// Unexpected EOF for last chunk could be a valid case. Any other errors are definitely real.
         .          .   2599:		if err != nil && !(errors.Is(err, io.ErrUnexpectedEOF) && i == len(pIdxs)-1) {
         .          .   2600:			return errors.Wrapf(err, "read range for seq %d offset %x", seq, pIdx.offset)
         .          .   2601:		}
         .          .   2602:
         .          .   2603:		chunkDataLen, n := binary.Uvarint(cb)
         .          .   2604:		if n < 1 {
         .          .   2605:			return errors.New("reading chunk length failed")
         .          .   2606:		}
         .          .   2607:
         .          .   2608:		// Chunk length is n (number of bytes used to encode chunk data), 1 for chunk encoding and chunkDataLen for actual chunk data.
         .          .   2609:		// There is also crc32 after the chunk, but we ignore that.
         .          .   2610:		chunkLen = n + 1 + int(chunkDataLen)
         .          .   2611:		if chunkLen <= len(cb) {
   1791358    3689948   2612:			err = populateChunk(&(res[pIdx.seriesEntry].chks[pIdx.chunk]), rawChunk(cb[n:chunkLen]), aggrs, r.save)
         .          .   2613:			if err != nil {
         .          .   2614:				return errors.Wrap(err, "populate chunk")
         .          .   2615:			}
         .          .   2616:			r.stats.chunksTouched++
         .          .   2617:			r.stats.chunksTouchedSizeSum += int(chunkDataLen)
(pprof)
```

I don't quite understand why (maybe @bboreham could shed some ligth on this) but for some reason that escapes me the `rawChunk` helper type is being sent to the heap before passing it to `populateChunk`. 

To fix this, we've adapted `populateChunk` to receive a byte slice instead, doing the cast inside the function and thus avoiding the unnecessary allocation.

### The results

To validate this changes we've resorted to `BenchmarkBlockSeries` by running the following command:

```sh
go test -run=X -bench=BenchmarkBlockSeries -benchmem -count=10 -cpu=8 ./pkg/storegateway
```

Here's the comparison between the old and the new implementation:

```
name                                                          old time/op    new time/op    delta
BlockSeries/concurrency:_1,_query_sharding_enabled:_false-8     7.82ms ± 4%    6.59ms ± 2%  -15.81%  (p=0.000 n=10+10)
BlockSeries/concurrency:_1,_query_sharding_enabled:_true-8      2.21ms ± 0%    2.07ms ± 1%   -6.31%  (p=0.000 n=10+9)
BlockSeries/concurrency:_2,_query_sharding_enabled:_false-8     5.06ms ± 1%    4.21ms ± 1%  -16.75%  (p=0.000 n=10+10)
BlockSeries/concurrency:_2,_query_sharding_enabled:_true-8      1.46ms ± 2%    1.32ms ± 1%   -9.62%  (p=0.000 n=10+10)
BlockSeries/concurrency:_4,_query_sharding_enabled:_false-8     3.61ms ± 4%    3.01ms ± 1%  -16.64%  (p=0.000 n=10+9)
BlockSeries/concurrency:_4,_query_sharding_enabled:_true-8      1.14ms ± 1%    1.00ms ± 1%  -12.60%  (p=0.000 n=9+10)
BlockSeries/concurrency:_8,_query_sharding_enabled:_false-8     3.70ms ±13%    3.18ms ± 3%  -13.93%  (p=0.000 n=10+9)
BlockSeries/concurrency:_8,_query_sharding_enabled:_true-8      1.10ms ± 1%    0.95ms ± 1%  -13.50%  (p=0.000 n=10+10)
BlockSeries/concurrency:_16,_query_sharding_enabled:_false-8    3.47ms ±11%    3.36ms ± 7%     ~     (p=0.218 n=10+10)
BlockSeries/concurrency:_16,_query_sharding_enabled:_true-8     1.05ms ± 1%    0.95ms ± 3%   -9.88%  (p=0.000 n=10+10)
BlockSeries/concurrency:_32,_query_sharding_enabled:_false-8    5.56ms ± 5%    4.99ms ±11%  -10.12%  (p=0.000 n=10+10)
BlockSeries/concurrency:_32,_query_sharding_enabled:_true-8     1.02ms ± 2%    0.92ms ± 3%   -9.88%  (p=0.000 n=9+10)

name                                                          old alloc/op   new alloc/op   delta
BlockSeries/concurrency:_1,_query_sharding_enabled:_false-8     17.7MB ± 2%    17.1MB ± 2%   -3.55%  (p=0.000 n=10+10)
BlockSeries/concurrency:_1,_query_sharding_enabled:_true-8      3.82MB ± 0%    3.76MB ± 0%   -1.82%  (p=0.000 n=10+9)
BlockSeries/concurrency:_2,_query_sharding_enabled:_false-8     18.3MB ± 1%    17.5MB ± 1%   -4.71%  (p=0.000 n=10+10)
BlockSeries/concurrency:_2,_query_sharding_enabled:_true-8      3.82MB ± 0%    3.82MB ± 0%     ~     (p=0.971 n=10+10)
BlockSeries/concurrency:_4,_query_sharding_enabled:_false-8     18.4MB ± 2%    17.8MB ± 1%   -3.27%  (p=0.000 n=10+10)
BlockSeries/concurrency:_4,_query_sharding_enabled:_true-8      3.82MB ± 0%    3.78MB ± 0%   -1.09%  (p=0.000 n=10+10)
BlockSeries/concurrency:_8,_query_sharding_enabled:_false-8     19.0MB ±13%    18.8MB ± 2%     ~     (p=0.211 n=10+9)
BlockSeries/concurrency:_8,_query_sharding_enabled:_true-8      3.81MB ± 0%    3.73MB ± 0%   -2.07%  (p=0.000 n=10+10)
BlockSeries/concurrency:_16,_query_sharding_enabled:_false-8    18.1MB ±12%    19.9MB ± 4%  +10.39%  (p=0.000 n=10+10)
BlockSeries/concurrency:_16,_query_sharding_enabled:_true-8     3.79MB ± 1%    3.72MB ± 1%   -1.72%  (p=0.000 n=10+10)
BlockSeries/concurrency:_32,_query_sharding_enabled:_false-8    31.9MB ± 6%    30.4MB ± 7%   -4.63%  (p=0.011 n=10+10)
BlockSeries/concurrency:_32,_query_sharding_enabled:_true-8     3.79MB ± 2%    3.70MB ± 2%   -2.39%  (p=0.000 n=10+10)

name                                                          old allocs/op  new allocs/op  delta
BlockSeries/concurrency:_1,_query_sharding_enabled:_false-8      58.0k ± 2%     30.3k ± 2%  -47.84%  (p=0.000 n=10+10)
BlockSeries/concurrency:_1,_query_sharding_enabled:_true-8       10.5k ± 0%      5.7k ± 0%  -45.79%  (p=0.000 n=9+10)
BlockSeries/concurrency:_2,_query_sharding_enabled:_false-8      60.1k ± 1%     30.9k ± 1%  -48.60%  (p=0.000 n=10+10)
BlockSeries/concurrency:_2,_query_sharding_enabled:_true-8       10.5k ± 0%      5.7k ± 0%  -45.79%  (p=0.000 n=10+10)
BlockSeries/concurrency:_4,_query_sharding_enabled:_false-8      60.4k ± 2%     31.5k ± 1%  -47.74%  (p=0.000 n=10+10)
BlockSeries/concurrency:_4,_query_sharding_enabled:_true-8       10.5k ± 0%      5.7k ± 0%  -45.86%  (p=0.000 n=10+10)
BlockSeries/concurrency:_8,_query_sharding_enabled:_false-8      62.5k ±13%     33.5k ± 2%  -46.49%  (p=0.000 n=10+9)
BlockSeries/concurrency:_8,_query_sharding_enabled:_true-8       10.5k ± 0%      5.7k ± 0%  -45.79%  (p=0.000 n=10+10)
BlockSeries/concurrency:_16,_query_sharding_enabled:_false-8     59.2k ±13%     35.4k ± 4%  -40.12%  (p=0.000 n=10+10)
BlockSeries/concurrency:_16,_query_sharding_enabled:_true-8      10.4k ± 1%      5.7k ± 1%  -45.15%  (p=0.000 n=10+10)
BlockSeries/concurrency:_32,_query_sharding_enabled:_false-8      106k ± 6%       54k ± 7%  -48.60%  (p=0.000 n=10+10)
BlockSeries/concurrency:_32,_query_sharding_enabled:_true-8      10.5k ± 2%      5.7k ± 2%  -45.40%  (p=0.000 n=10+10)
```

According to this, with the new implementation the number of total allocations is reduced by a `~90%`.

~~Worth mentioning, that operation spent time seems to have worsened. However, I think it's fair to attribute this to noise, as times barely differ in `~1ms`.~~ See discussion: https://github.com/grafana/mimir/pull/3215#issuecomment-1280660724

#### Which issue(s) this PR fixes or relates to

Fixes #3306 

#### Checklist

- [X] Tests updated
- [ ] Documentation added
- [X] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`