@colega 2022-10-17
> Worth mentioning, that operation spent time seems to have worsened. However, I think it's fair to attribute this to noise, as times barely differ in ~1ms.

I think that millisecond scale is not as neglible as it can be micro/nanosecond, can we try understand where that increase comes from? 
One of the things I'd blame is that async put into pool.