@ortuman 2022-10-17
> I think that millisecond scale is not as neglible as it can be micro/nanosecond, can we try understand where that increase comes from?
One of the things I'd blame is that async put into pool.

Got some CPU profile from the benchmark, and it seems that a significant amount of time is now spent in the pool `Put` method. 

This makes sense, given that in some benchmark scenarios the total number of returned chunks is around **500k**. 😱  

![Screenshot 2022-10-17 at 15 22 18](https://user-images.githubusercontent.com/888899/196191691-8482fffd-ae3e-4b93-97ec-34896c418bf0.png)

Maybe establishing some max returned objects limit here is not a bad option after all. wdyt?