@ortuman 2022-10-18
In the end, I decided to adapt chunks pool to contain slices with a total capacity of `10k` chunks per slice, rather than managing plain `storepb.Chunk` objects.

And now benchmark results look better (see the updated comparison in the PR description).

This reduces drastically total required pool invocations, so I've also decided to remove the async method.