@colega 2022-10-18
That's an interesting approach, where did get the 10k number? Or is it just a _big but not excessively huge_ approach?