@ortuman 2022-10-18
> Or is it just a big but not excessively huge approach?

That's the idea. Though still an arbitrary number, susceptible to being changed.

I'll add some comment to clarify. 👍