@colega 2022-10-20
The changes LGTM, I'd still like to discuss whether it makes sense to have those pools per-block or they should be global, let's see if @pracucci & @bboreham can say something on that.